requirements.txt
Flask==3.0.1
gunicorn==21.2.0
peewee==3.17.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1

Flask-Cors==4.0.0
